Move code to db.go
